Move helper classes to proper namespace
